fix(#6137): split Winsock and CStdLib syscall tests out of SyscallTest.java
SyscallTest.java sat four lines below the 1000-line FileLengthCheck
ceiling, with no room for the next test that touches sockets - #6133
hit exactly that. The length came from WindowsSocketTest and
PosixSocketTest, the same seven cases written twice over, differing
only in which native binding they call.

Move each half next to the production class it drives: WinsockTest
sits beside Winsock.java, CStdLibTest beside CStdLib.java, satisfying
jtcop's rule that a test class must sit next to a real production
class. RandomServer and RandomPort become small shared fixtures both
halves and SyscallTest use. SyscallTest keeps only the cases that
drive the EO socket object, at 248 lines.

Update the simian exclude to name the two new files instead of
SyscallTest.java, matching the intent behind the earlier #6113 fix.
