+++ .github/workflows/simian.yml
@@ -21,4 +21,4 @@ jobs:
           distribution: 'temurin'
           java-version: 17
       - run: wget --quiet https://public.yegor256.com/simian.jar -O /tmp/simian.jar
-      - run: java -jar /tmp/simian.jar -threshold=15 "-excludes=**/SyscallTest.java" "-excludes=**/gen" "-excludes=**/it" "**/*.java"
+      - run: java -jar /tmp/simian.jar -threshold=15 "-excludes=**/WinsockTest.java" "-excludes=**/CStdLibTest.java" "-excludes=**/gen" "-excludes=**/it" "**/*.java"
